requirements.txt
requests
lxml
python_telegram_bot
nose
